+++ spec/controllers/location_commands_controller_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe LocationCommandsController, type: :controller do
 
     context 'with non-UUID id' do
       [1, 'fo', '1', SecureRandom.uuid[1..-4]].each do |id|
-        fit "returns 400 for id=#{id}" do
+        it "returns 400 for id=#{id}" do
           put :create_by_id, params: { id: id, name: name }
 
           expect(response).to have_http_status(:bad_request)