+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed:
+- `location_queries/inside` returns 400 Bad Request instead of 500 when given query is not a Point geometry.
+
 ## [1.1.0] - 2019-11-27
 ### Changed:
 - Use own implementation of point-in-polygon check; this removes runtime dependency on geos.