+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - `location_queries/inside` returns 400 Bad Request instead of 500 when given query is not a Point geometry.
 
+### Added
+- `PUT location_commands/create_by_id` endpoint for creating new Location with id - idempotent,
+- `POST location_commands/create` endpoint for creating new Location, non-idempotent,
+
 ## [1.1.0] - 2019-11-27
 ### Changed:
 - Use own implementation of point-in-polygon check; this removes runtime dependency on geos.