+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.0] - 2019-11-29
 ### Changed:
 - `location_queries/inside` returns 400 Bad Request instead of 500 when given query is not a Point geometry.
 
@@ -29,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   standardized open source project CHANGELOG.
 - Skeleton of the new microservice with rails & rspec
 
-[Unreleased]: https://github.com/scoiatael/medea/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/scoiatael/medea/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/scoiatael/medea/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/scoiatael/medea/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/scoiatael/medea/compare/v0.0.1...v1.0.0
 [0.0.1]: https://github.com/scoiatael/medea/releases/tag/v0.0.1